+++ app/javascript/app.vue
@@ -1,10 +1,8 @@
 <template>
   <div>
-    <TheHeader></TheHeader>
-    <div>
-      <p>{{ message }}</p>
-    </div>
-    <TheFooter></TheFooter>
+    <TheHeader />
+    <router-view />
+    <TheFooter />
   </div>
 </template>
 
@@ -16,17 +14,5 @@ export default {
     TheFooter, 
     TheHeader 
   },
-  data: function () {
-    return {
-      message: "Hello Vue!"
-    }
-  }
 }
 </script>
-
-<style scoped>
-p {
-  font-size: 2em;
-  text-align: center;
-}
-</style>
