+++ app/javascript/packs/hello_vue.js
@@ -7,10 +7,12 @@
 
 import Vue from 'vue'
 import App from '../app.vue'
+import router from '../router'
 import 'bootstrap/dist/css/bootstrap.css'
 
 document.addEventListener('DOMContentLoaded', () => {
   const app = new Vue({
+    router,
     render: h => h(App)
   }).$mount()
   document.body.appendChild(app.$el)
