+++ config/routes.rb
@@ -1,3 +1,4 @@
 Rails.application.routes.draw do
   root to: 'home#index'
+  get '*path', to: 'home#index'
 end
