+++ package.json
@@ -10,6 +10,7 @@
     "turbolinks": "^5.2.0",
     "vue": "^2.6.14",
     "vue-loader": "^15.9.8",
+    "vue-router": "^3.5.2",
     "vue-template-compiler": "^2.6.14"
   },
   "version": "0.1.0",
